Image display, TagLib is now used for metadata...
Cover art can now be displayed for songs with an ID3(v2) tag thanks
to TagLib. This does, however, mean that the user must have TagLib
installed. At some later point, as TagLib is under the LGPL-2.1
license, it might be added as a dependency in the project itself,
instead of requiring that it be installed first.

The layout has changed a bit, and will likely still be changed further
in the future.

The player and volume controls have been modified to use icons from
https://icons8.com, which should hopefully be free to use as the
project is open source.
The volume button now changes it's icon when the volume changes, so
it has 4 different stages (Mute, Low, Medium, High) that it cycles
through.

I'm still working on playlist's and how songs are played in general,
as well as what data is displayed on screen.